requirements.txt
matplotlib==3.8.1
matplotlib_label_lines==0.7.0
mpl_scatter_density==0.7
numpy==2.0.1
pandas==2.2.2
scipy==1.14.0
seaborn==0.13.2
tqdm==4.66.1
